MCR-3067 move test helper methods to test classes
